requirements.txt
flask==2.2.2
flask-cors==3.0.10
flask-restful==0.3.9
psycopg2-binary==2.8.6
